[BB-590] fix client ListOnCallUsers schedule id param (#10)
#### Description
The rootly api for `/v1/shifts` expects the schedule ids query param to be `&schedule_ids[]=your-schedule-id` not `&schedule_ids=[your-schedule-id]`. This change updates the rootly client to former. Unfortunately with the latter format rootly returns shifts for all schedules, instead of responding back with no matches or an error.